Consolidate Admin: tabs + shared layout (#231)
Move /admin/clients, /admin/comps, /admin/evaluations, /admin/feedback, and
the current /admin (Invite Codes + Plugin-beta Users) into a single tabbed
admin shell:

- New src/app/admin/layout.tsx renders the "Admin" header, the platform-admin
  gate (one fetch of /api/admin/status), and the tab bar across every admin
  child route. Replaces the four button-style links that used to sit on
  /admin's top-right.
- Tab order: Clients (default) / Evaluations / Feedback / Comps / Beta Codes.
- Extract TabButton into src/components/Tabs.tsx, reused by the layout and by
  the team member detail page (which previously defined its own local copy).
- /admin redirects to /admin/clients.
- Move the current /admin content to /admin/beta-codes (the page is really a
  beta-codes tool).
- Each existing admin sub-page loses its own pt-20 wrapper, its own page h1,
  and its own authorized + access-required panel, since the layout owns all
  three. Section content unchanged.
- Account menu collapses the two admin entries ("Clients" + "Admin") into one
  "Admin" row that lands on the Clients tab.

Behavior, data, and auth result are unchanged: the layout's admin gate
produces the same access decision the per-page checks did, and the existing
sub-routes still respond at the same URLs.

/hq journeys + features updated in lockstep (lastPr to be corrected once the
PR number is known).
