+++ src/components/UserMenu.tsx
@@ -345,15 +345,11 @@ export function UserMenu() {
             />
           </div>
 
-          {/* Admin (platform admins only) */}
+          {/* Admin (platform admins only). One unified entry — the Admin page
+              now hosts Clients, Evaluations, Feedback, Comps, and Beta Codes
+              as tabs (#231). Lands on Clients by default. */}
           {isAdmin && (
             <div className="py-1.5 border-t border-[#2a2a2a]">
-              <MenuRow
-                icon={ICON.clients}
-                label="Clients"
-                href="/admin/clients"
-                onClick={() => setOpen(false)}
-              />
               <MenuRow
                 icon={ICON.admin}
                 label="Admin"
